requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
pymupdf
python-dotenv
tqdm
requests